+++ src/game/g_match.c
@@ -502,7 +502,7 @@ char *G_createStats(gentity_t *refEnt)
 		}
 	}
 
-	return(va("%d %d %d%s %d%s", refEnt - g_entities,
+	return(va("%d %d %d%s %d%s", (int)(refEnt - g_entities),
 	          refEnt->client->sess.rounds,
 	          dwWeaponMask,
 	          strWeapInfo,
