+++ src/game/g_systemmsg.c
@@ -130,7 +130,7 @@ void G_SendSystemMessage(sysMsg_t message, int team)
 			continue;
 		}
 
-		trap_SendServerCommand(other - g_entities, va("vschat 0 %d 3 %s 0 0 0", other - g_entities, systemMessages[message]));
+		trap_SendServerCommand(other - g_entities, va("vschat 0 %d 3 %s 0 0 0", (int)(other - g_entities), systemMessages[message]));
 	}
 }
